audio: bump volume on DEC7 and DEC9
Some applications (mainly recording apps) use one or both of the digital microphones. As it passes trough the DEC7 or DEC9 Multiplexer it encodes it at a too low volume, making the
recorded playback too quiet. Bumping the DEC7 and DEC9 volume to 200 is only a workaround and can most likely be changed in the kernel

Change-Id: Ifa4f9fc44f38537d6cce683fef4b9e463627700b
Signed-off-by: CTXz <ctx.xda@gmail.com>